refactor: move Skeleton and AudioPlayerSkeleton
