+++ apps/rtk-query/src/shared/components/AudioPlayer/AudioPlayerSceleton/AudioPlayerSkeleton.tsx
