+++ apps/rtk-query/src/shared/components/Skeleton/Skeleton.module.css
