+++ apps/rtk-query/src/shared/components/Skeleton/Skeleton.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react-vite'
 
-import { Card } from '../../Card'
-import { Typography } from '../../Typography'
+import { Card } from '../Card'
+import { Typography } from '../Typography'
 import { Skeleton } from './Skeleton.tsx'
 
 const meta = {
