+++ apps/rtk-query/src/shared/components/Skeleton/Skeleton.tsx
