+++ apps/rtk-query/src/shared/components/Skeleton/index.ts
@@ -1,1 +1,1 @@
-export * from '../AudioPlayer/AudioPlayerSceleton/Skeleton.tsx'
+export * from './Skeleton.tsx'
